+++ vitest.config.ts
@@ -1,5 +1,4 @@
-import { defineConfig, mergeConfig } from 'vitest/config';
-import type { ViteUserConfig } from 'vitest/config';
+import { defineConfig, mergeConfig, type ViteUserConfig } from 'vitest/config';
 
 const configExtension: { [key: string]: ViteUserConfig } = {
   unit: defineConfig({
@@ -22,7 +21,7 @@ const configExtension: { [key: string]: ViteUserConfig } = {
         ],
         thresholds: {
           lines: 79,
-          functions: 83,
+          functions: 82,
           statements: 78,
           branches: 71,
         },
@@ -47,6 +46,7 @@ export default mergeConfig(
     test: {
       globals: true,
       restoreMocks: true,
+      mockReset: true,
       environment: 'node',
       env: {
         FORCE_COLOR: '1',